+++ src/main/resources/templates/searchPage.html
@@ -7,7 +7,7 @@
 <body>
 <th:block th:include="/_menu"></th:block>
 
-<form name='f' action="#" th:action="@{/searchingResult(query=${query})}"  method="get">
+<form name='f' action="#" th:action="@{/search(query=${query})}"  method="get">
     <input name="query" placeholder="Type query...">
     <button type="submit">Find</button>
 </form>
